requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
bcrypt
python-multipart
psutil
apscheduler
requests
